ej 3 métodos

+++ src/main/java/com/mireyaserrano/tema04/Ejercicio3.java
@@ -3,8 +3,8 @@ package com.mireyaserrano.tema04;
 import java.util.Scanner;
 
 public class Ejercicio3 {
-    public static Float mayorNum(float num1, float num2){
-        float mayor;
+    public static int mayorNum(int num1, int num2){
+        int mayor;
         if (num1>num2){
             mayor = num1;
         }else {
@@ -16,9 +16,9 @@ public class Ejercicio3 {
     public static void main(String args[]){
         Scanner scanner = new Scanner(System.in);
         System.out.println("introduce el primer número");
-        float x = Float.parseFloat(scanner.nextLine());
+        int x = Integer.parseInt(scanner.nextLine());
         System.out.println("introduce el segundo número");
-        float y = Float.parseFloat(scanner.nextLine());
+        int y = Integer.parseInt(scanner.nextLine());
         scanner.close();
         System.out.println("el número mayor es " + mayorNum(x,y));
     }
